Bugfixes und neue Bugs
Alte Bugs gefixt
Neuer externer Fehler, ka wie entsteht

+++ ki/ki/KategorieMitJahrenUndWerten.cs
@@ -20,21 +20,21 @@ namespace ki
             this.category = category;
             YearsWithValues = YearsWithValuesTask.Result;
         }
-     public   static List<double> GetYearsFromList(List<YearWithValue> list)
+     public   static List<float> GetYearsFromList(List<YearWithValue> list)
         {
-            List<double> jahre = new List<double>();
+            List<float> jahre = new List<float>();
             foreach (var item in list)
             {
-                jahre.Add(item.Year);
+                jahre.Add((float)item.Year);
             }
             return jahre;
         }
-        public static List<decimal> GetValuesFromList(List<YearWithValue> list)
+        public static List<float> GetValuesFromList(List<YearWithValue> list)
         {
-            List<decimal> werte = new List<decimal>();
+            List<float> werte = new List<float>();
             foreach (var item in list)
             {
-                werte.Add(item.Value);
+                werte.Add((float)(item.Value));
             }
          
             return werte;
